Run FastQC

This function calls fastqc using system2(). To run in parallel, register a parallel backend, e.g., using doParallel::registerDoParallel().

Usage

fastqc(filepaths, outputDir = "fastqc_output", cmd = "fastqc", args = NULL)

Arguments

filepaths

Paths to fastq files. For single-end reads, each element should be a single filepath. For paired-end reads, each element can be two filepaths separated by ";".

outputDir

Directory in which to store output. Will be created if it doesn't exist.

cmd

Name or path of the command-line interface.

args

Additional arguments to pass to the command-line interface.

Value

A vector of exit codes, invisibly.

See also