Install seeker's system dependencies

This function installs and configures the various programs required for seeker to fetch and process RNA-seq data.

Usage

installSysDeps(
  sraToolkitDir,
  minicondaDir,
  refgenieDir,
  rprofileDir,
  minicondaEnv = "seeker",
  refgenieGenomes = NULL,
  fastqscreenDir = NULL
)

Arguments

sraToolkitDir

String indicating directory in which to install the SRA Toolkit. Recommended to use "~", the home directory. If NULL, the Toolkit will not be installed.

minicondaDir

String indicating directory in which to install Miniconda. Recommended to use "~", the home directory. If NULL, Miniconda will not be installed.

refgenieDir

String indicating directory in which to store the directory of genome assets from refgenie, which will be named "refgenie_genomes". Recommended to use "~", the home directory. Only used if minicondaDir is not NULL.

rprofileDir

String indicating directory in which to create or modify .Rprofile, which is run by R on startup. Common options are "~" or ".".

minicondaEnv

String indicating name of the Miniconda environment in which to install various conda packages (fastq-screen, fastqc, multiqc, pigz, refgenie, salmon, and trim-galore).

refgenieGenomes

Character vector indicating genome assets, such as transcriptome indexes for salmon(), to pull from refgenomes using refgenie. If NULL, no assets are fetched.

fastqscreenDir

String indicating directory in which to download the genomes for fastqscreen(). This takes a long time. If NULL, genomes are not downloaded.

Value

NULL, invisibly

See also