Run Trim Galore!

This function calls trim_galore using system2(), and is only designed to handle standard adapter/quality trimming. To run in parallel, register a parallel backend, e.g., using doParallel::registerDoParallel().

Usage

trimgalore(
  filepaths,
  outputDir = "trimgalore_output",
  cmd = "trim_galore",
  args = NULL,
  pigzCmd = "pigz"
)

Arguments

filepaths

Paths to fastq files. For single-end reads, each element should be a single filepath. For paired-end reads, each element should be two filepaths separated by ";".

outputDir

Directory in which to store output. Will be created if it doesn't exist.

cmd

Name or path of the command-line interface.

args

Additional arguments to pass to the command-line interface. Output files will always be compressed. Arguments "–gzip", "–cores", "-j", and "–basename" are not allowed. Arguments "-o" and "–paired" should not be specified here.

pigzCmd

String for pigz command, which will gzip the output files.

Value

A vector of exit codes, invisibly.

See also