Run tximport on RNA-seq quantifications

This function uses the tximport package.

Usage

tximport(
  inputDir,
  tx2gene,
  samples = NULL,
  outputDir = "data",
  type = c("salmon", "kallisto"),
  countsFromAbundance = "lengthScaledTPM",
  ignoreTxVersion = TRUE,
  ...
)

Arguments

inputDir

Directory that contains the quantification directories.

tx2gene

NULL or data.frame of mapping between transcripts and genes, as returned by getTx2gene(), passed to tximport::tximport().

samples

Names of quantification directories to include. NULL indicates all.

outputDir

Directory in which to save the result, a file named "tximport_output.qs", using qs::qsave(). If NULL, no file is saved.

type

Passed to tximport::tximport().

countsFromAbundance

Passed to tximport::tximport().

ignoreTxVersion

Passed to tximport::tximport().

...

Additional arguments passed to tximport::tximport().

Value

A list, as returned by tximport::tximport(), invisibly.

See also