Process microarray data end to end

This function fetches data and metadata from NCBI GEO and ArrayExpress, processes raw Affymetrix data using RMA and custom CDFs from Brainarray, and maps probes to genes.

seekerArray(params, parentDir)

Arguments

params

Named list of parameters with components:

study: String indicating the study accession and used to name the output directory within parentDir. Must start with "GSE", "E-", or "LOCAL". If starts with "GSE", data are fetched using GEOquery::getGEO(). If starts with "E-", data are fetched using ArrayExpress::getAE(). If starts with "LOCAL", data in the form of cel(.gz) files must in the directory parentDir/params$study/raw, and parentDir/params$study must contain a file "sample_metadata.csv" that has a column sample_id containing the names of the cel(.gz) files without the file extension.

geneIdType: String indicating whether to map probes to gene IDs from Ensembl ("ensembl") or Entrez ("entrez").

platform: String indicating the GEO-based platform accession for the raw data. See https://www.ncbi.nlm.nih.gov/geo/browse/?view=platforms. Only necessary if params$study starts with "LOCAL", or starts with "GSE" and the study uses multiple platforms.

params can be derived from a yaml file, see vignette("introduction", package = "seeker"). The yaml representation of params will be saved to parentDir/params$study/params.yml.

parentDir

Directory in which to store the output, which will be a directory named according to params$study.

Value

NULL, invisibly.

Details

The standard output:

naive_expression_set.qs: Initial ExpresssionSet generated by GEOquery::getGEO or ArrayExpress::ae2bioc(). Should generally not be used if sample_metadata.csv and gene_expression_matrix.qs are available.

sample_metadata.csv: Table of sample metadata. Column sample_id matches colnames of the gene expression matrix.

gene_expression_matrix.qs: Rows correspond to genes, columns to samples.

custom_cdf_name.txt: Name of custom CDF package used by affy::justRMA() to process and normalize raw Affymetrix data and map probes to genes.

feature_metadata.qs: GPL object, if gene expression matrix was generated from processed data.

probe_gene_mapping.csv.gz: Table of probes and genes, if gene expression matrix was generated from processed data.

"raw" directory: Contains raw Affymetrix files.

params.yml: Parameters used to process the dataset.

session.log: R session information.

The output may include other files from NCBI GEO or ArrayExpress. Files with extension "qs" can be read into R using qs::qread().